Make KotlinFX work with Kotlin M8. Closes #3
It had mainly something to do with the new visibility changes of the
Kotlin compiler. I hade to make the layout classes ending with a
'K' public. That is actually not good though because they are not
meant to be public. Maybe in a future Kotlin release it will work.